requirements.txt
streamlit
transformers
torch
sentencepiece
python-dotenv
pyttsx3
